install fixes set fake season to 2016 make install script take name of DB as argument

+++ INSTALL.debian.sh
@@ -1,24 +1,33 @@
 #!/bin/sh
 #apt-get install python-mysqldb  php5-mysqlnd mysql nodejs npm
 
-mysql -u roprotokol --password=roprotokol roprotokol<<EOSQL
-  DROP  DATABASE IF EXISTS roprotokol;
-  CREATE SCHEMA roprotokol DEFAULT CHARACTER SET utf8 COLLATE utf8_general_ci;
-  CREATE USER 'roprotokol'@'localhost' IDENTIFIED BY 'roprotokol';
-  GRANT ALL PRIVILEGES ON 'roprotokol'.* TO 'roprotokol'@'localhost';
-  FLUSH PRIVILEGES;
-EOSQL
 
+
+defaultDB=roprotokolx
+
+DB=${1-$defaultDB}
+
+echo using database $DB
+if [ ! -f backend/convert/secret.sh ]; then
+    echo 'DBCMD="mysql -u roprotokol --password=roprotokol ' $DB '"' >  backend/convert/secret.sh
+    echo 'superDBCMD="mysql -u roprotokol --password=roprotokol"' >  backend/convert/secret.sh
+    echo 'dbpassword=roprotokol' >>  backend/convert/secret.sh
+    echo "roprotokol" > backend/tests/secret.db    
+fi
+
+
+. backend/convert/secret.sh
+
+
+echo using DBCMD $DBCMD $superDBCMD
+echo "DROP DATABASE IF EXISTS $DB;"| $superDBCMD
+echo "CREATE DATABASE IF NOT EXISTS $DB;" | $superDBCMD
+echo created db
 
 echo NOW mkdb.sql
-mysql -u roprotokol -p'roprotokol' roprotokol < backend/convert/mkdb.sql
-# mysql -u roprotokol -p'roprotokol' roprotokol <  backend/convert/queries.sql
-
-
-echo 'DBCMD="mysql -u roprotokol --password=roprotokol roprotokol"' >  backend/convert/secret.sh
-echo "roprotokol" > backend/tests/secret.db
+$DBCMD < backend/convert/mkdb.sql
 
 echo NOW FAKE
-./backend/convert/import.sh fake
+./backend/convert/import.sh $DB fake $dbpassword
 
 echo now configure your webserver to server DSR roprotokol
